requirements.txt
anyio==3.6.2
asgiref==3.6.0
astroid==2.15.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
dill==0.3.6
ecdsa==0.18.0
fastapi==0.94.1
fastapi_keycloak==1.0.7
future==0.18.3
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
minio==7.1.13
platformdirs==3.1.1
pyArango==2.0.1
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.6
PyJWT==2.6.0
pylint==2.17.0
python-arango==7.5.7
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
zope.interface==6.0
